+++ src/main/java/com/zanitech/sdk/tcmbclient/TcmbClient.java
@@ -7,10 +7,13 @@ public class TcmbClient {
 
   private ForexAdapter forexAdapter = new ForexAdapterImpl();
 
+  /**
+   * Foreign exchange related functionalities
+   * @return {@link ForexAdapter}
+   */
   public ForexAdapter forex() {
     return forexAdapter;
   }
-
 
   public void setForexAdapter(ForexAdapter forexAdapter) {
     this.forexAdapter = forexAdapter;
