+++ src/main/java/com/zanitech/sdk/tcmbclient/adapter/ForexAdapter.java
@@ -1,8 +1,12 @@
 package com.zanitech.sdk.tcmbclient.adapter;
 
-import com.zanitech.sdk.tcmbclient.model.currency.ForexRateOutput;
+import com.zanitech.sdk.tcmbclient.model.forex.ForexRateOutput;
 
 public interface ForexAdapter {
 
+  /**
+   * Instant foreign exchange rates listed by TCMB
+   * @return {@link ForexRateOutput}
+   */
   ForexRateOutput getRates();
 }
