--- src/main/java/com/zanitech/sdk/tcmbclient/adapter/impl/BankAndBranchAdapterImpl.java
@@ -1,8 +1,0 @@
-package com.zanitech.sdk.tcmbclient.adapter.impl;
-
-public class BankAndBranchAdapterImpl {
-
-  public String getBank() {
-    return "Akbank T.A.Ş";
-  }
-}
