+++ src/test/java/com/zanitech/sdk/tcmbclient/adapter/ExhangeRateFakeAdapter.java
@@ -1,7 +1,7 @@
 package com.zanitech.sdk.tcmbclient.adapter;
 
-import com.zanitech.sdk.tcmbclient.model.currency.ForexRate;
-import com.zanitech.sdk.tcmbclient.model.currency.ForexRateOutput;
+import com.zanitech.sdk.tcmbclient.model.forex.ForexRate;
+import com.zanitech.sdk.tcmbclient.model.forex.ForexRateOutput;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
